refactor: extract propagate business logic from commands/ to core/refactor/propagate.rs
Move ~370 lines of struct field propagation logic (find_struct_definition,
extract_struct_source, apply_propagate_edits, walk_rs_files, PropagateField,
PropagateEdit types) from commands/refactor.rs into core/refactor/propagate.rs.

commands/refactor.rs run_propagate() is now a thin wrapper: resolve root,
capture undo snapshot, call core::propagate(), log results. Uses canonical
codebase_scan::walk_files instead of custom walk_rs_files/walk_rs_recursive.
